refactor(bot): move getDebugPlayers fixture to debugFixtures.ts (T15)
Extract the 13-row WoWPlayer fixture out of `core/utils.ts` (which is
otherwise Discord-member helpers) into its own `core/debugFixtures.ts`
module. Update GroupService and tests to import from the new path; move
the dedicated unit test alongside the fixture.
