+++ packages/bot/src/services/groupService.ts
@@ -1,6 +1,7 @@
 import { WoWGroup, WoWPlayer, createMythicPlusGroups, setGroupHistory, todayPST } from '@mythicplus/shared';
 import { announceGroup, type Sendable } from '../core/groupUi.js';
-import { getDebugPlayers, getPlayerList, type DiscordMember, type TypingChannel } from '../core/utils.js';
+import { getPlayerList, type DiscordMember, type TypingChannel } from '../core/utils.js';
+import { getDebugPlayers } from '../core/debugFixtures.js';
 import { FirebaseService } from '../core/firebaseService.js';
 import logger from '../core/logger.js';
